# perf: Batch parquet primitive decoding