--- a/crates/polars-arrow/src/bitmap/utils/iterator.rs
+++ b/crates/polars-arrow/src/bitmap/utils/iterator.rs
@@ -46,22 +46,66 @@ impl<'a> BitmapIter<'a> {
             rest_len,
         }
     }
-}
 
-impl<'a> Iterator for BitmapIter<'a> {
-    type Item = bool;
+    /// Consume and returns the numbers of `1` / `true` values at the beginning of the iterator.
+    ///
+    /// This performs the same operation as `(&mut iter).take_while(|b| b).count()`.
+    ///
+    /// This is a lot more efficient than consecutively polling the iterator and should therefore
+    /// be preferred, if the use-case allows for it.
+    pub fn take_leading_ones(&mut self) -> usize {
+        let word_ones = usize::min(self.word_len, self.word.trailing_ones() as usize);
+        self.word_len -= word_ones;
+        self.word = self.word.wrapping_shr(word_ones as u32);
 
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
         if self.word_len != 0 {
-            let ret = self.word & 1 != 0;
-            self.word >>= 1;
-            self.word_len -= 1;
-            return Some(ret);
+            return word_ones;
         }
 
-        if self.rest_len != 0 {
-            self.word_len = self.rest_len.min(64);
+        let mut num_leading_ones = word_ones;
+
+        while self.rest_len != 0 {
+            self.word_len = usize::min(self.rest_len, 64);
+            self.rest_len -= self.word_len;
+
+            unsafe {
+                let chunk = self.bytes.get_unchecked(..8).try_into().unwrap();
+                self.word = u64::from_le_bytes(chunk);
+                self.bytes = self.bytes.get_unchecked(8..);
+            }
+
+            let word_ones = usize::min(self.word_len, self.word.trailing_ones() as usize);
+            self.word_len -= word_ones;
+            self.word = self.word.wrapping_shr(word_ones as u32);
+            num_leading_ones += word_ones;
+
+            if self.word_len != 0 {
+                return num_leading_ones;
+            }
+        }
+
+        num_leading_ones
+    }
+
+    /// Consume and returns the numbers of `0` / `false` values that the start of the iterator.
+    ///
+    /// This performs the same operation as `(&mut iter).take_while(|b| !b).count()`.
+    ///
+    /// This is a lot more efficient than consecutively polling the iterator and should therefore
+    /// be preferred, if the use-case allows for it.
+    pub fn take_leading_zeros(&mut self) -> usize {
+        let word_zeros = usize::min(self.word_len, self.word.trailing_zeros() as usize);
+        self.word_len -= word_zeros;
+        self.word = self.word.wrapping_shr(word_zeros as u32);
+
+        if self.word_len != 0 {
+            return word_zeros;
+        }
+
+        let mut num_leading_zeros = word_zeros;
+
+        while self.rest_len != 0 {
+            self.word_len = usize::min(self.rest_len, 64);
             self.rest_len -= self.word_len;
             unsafe {
                 let chunk = self.bytes.get_unchecked(..8).try_into().unwrap();
@@ -69,19 +113,56 @@ impl<'a> Iterator for BitmapIter<'a> {
                 self.bytes = self.bytes.get_unchecked(8..);
             }
 
-            let ret = self.word & 1 != 0;
-            self.word >>= 1;
-            self.word_len -= 1;
-            return Some(ret);
+            let word_zeros = usize::min(self.word_len, self.word.trailing_zeros() as usize);
+            self.word_len -= word_zeros;
+            self.word = self.word.wrapping_shr(word_zeros as u32);
+            num_leading_zeros += word_zeros;
+
+            if self.word_len != 0 {
+                return num_leading_zeros;
+            }
         }
 
-        None
+        num_leading_zeros
+    }
+
+    /// Returns the number of remaining elements in the iterator
+    #[inline]
+    pub fn num_remaining(&self) -> usize {
+        self.word_len + self.rest_len
+    }
+}
+
+impl<'a> Iterator for BitmapIter<'a> {
+    type Item = bool;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.word_len == 0 {
+            if self.rest_len == 0 {
+                return None;
+            }
+
+            self.word_len = self.rest_len.min(64);
+            self.rest_len -= self.word_len;
+
+            unsafe {
+                let chunk = self.bytes.get_unchecked(..8).try_into().unwrap();
+                self.word = u64::from_le_bytes(chunk);
+                self.bytes = self.bytes.get_unchecked(8..);
+            }
+        }
+
+        let ret = self.word & 1 != 0;
+        self.word >>= 1;
+        self.word_len -= 1;
+        Some(ret)
     }
 
     #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
-        let exact = self.word_len + self.rest_len;
-        (exact, Some(exact))
+        let num_remaining = self.num_remaining();
+        (num_remaining, Some(num_remaining))
     }
 }
 
@@ -102,3 +183,59 @@ impl<'a> DoubleEndedIterator for BitmapIter<'a> {
 
 unsafe impl TrustedLen for BitmapIter<'_> {}
 impl ExactSizeIterator for BitmapIter<'_> {}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    #[ignore = "Fuzz test. Too slow"]
+    fn test_leading_ops() {
+        for _ in 0..10_000 {
+            let bs = rand::random::<u8>() % 4;
+
+            let mut length = 0;
+            let mut pattern = Vec::new();
+            for _ in 0..rand::random::<usize>() % 1024 {
+                let word = match bs {
+                    0 => u64::MIN,
+                    1 => u64::MAX,
+                    2 | 3 => rand::random(),
+                    _ => unreachable!(),
+                };
+
+                pattern.extend_from_slice(&word.to_le_bytes());
+                length += 64;
+            }
+
+            for _ in 0..rand::random::<usize>() % 7 {
+                pattern.push(rand::random::<u8>());
+                length += 8;
+            }
+
+            let last_length = rand::random::<usize>() % 8;
+            if last_length != 0 {
+                pattern.push(rand::random::<u8>());
+                length += last_length;
+            }
+
+            let mut iter = BitmapIter::new(&pattern, 0, length);
+
+            let mut prev_remaining = iter.num_remaining();
+            while iter.num_remaining() != 0 {
+                let num_ones = iter.clone().take_leading_ones();
+                assert_eq!(num_ones, (&mut iter).take_while(|&b| b).count());
+
+                let num_zeros = iter.clone().take_leading_zeros();
+                assert_eq!(num_zeros, (&mut iter).take_while(|&b| !b).count());
+
+                // Ensure that we are making progress
+                assert!(iter.num_remaining() < prev_remaining);
+                prev_remaining = iter.num_remaining();
+            }
+
+            assert_eq!(iter.take_leading_zeros(), 0);
+            assert_eq!(iter.take_leading_ones(), 0);
+        }
+    }
+}