… crates/polars-parquet/src/arrow/read/deserialize/primitive/basic.rs
… crates/polars-parquet/src/arrow/read/deserialize/utils.rs
… crates/polars-parquet/src/parquet/encoding/hybrid_rle/buffered.rs
… crates/polars-parquet/src/parquet/encoding/hybrid_rle/fuzz.rs
… crates/polars-parquet/src/parquet/encoding/hybrid_rle/mod.rs
… crates/polars-parquet/src/parquet/encoding/hybrid_rle/translator.rs
… crates/polars/tests/it/io/parquet/read/binary.rs
… crates/polars/tests/it/io/parquet/read/fixed_binary.rs
… crates/polars/tests/it/io/parquet/read/primitive.rs
… crates/polars/tests/it/io/parquet/read/primitive_nested.rs
… crates/polars/tests/it/io/parquet/read/struct_.rs
… crates/polars/tests/it/io/parquet/read/utils.rs